clk: qcom: clk-cpu-osm: Query current CPU frequency from OSM
Instead of getting the index and consulting the in-kernel frequency
table for the frequency corresponding to the fetched index, read the XO
clock multiplier for the current frequency directly from OSM and
multiply it by the XO rate. This ensures that if the kernel frequency
table has been altered, accurate readings can still be obtained.

To prevent cpufreq from thinking the frequencies are out of sync when
a CPU is on the 300 Mhz step, detect this specific situation and report
300 MHz. Without special care, the reported frequency would be 288 MHz
as this is what the CPU is actually running at due to its multiplier of
15 and base XO rate of 19.2 MHz. Qualcomm appears to have rounded off
the frequency for brevity on the kernel side.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Papi64s <mhmmdsptra@outlook.com>